Update OpenMote board header file to define UPDATE_CCA.
As per the cc2538dk and ReMote boards. This enables the cc2538's
bootloader backdoor, to allow flashing via the serial interface even
when a valid image is already present.

Backdoor enable is set to LOW, PORT A, PIN 6 (ON/SLEEP on the OpenBase),
for consistency with the default OpenMote firmware.